split up css formats
